Refactored user Schema by adding contract class
