+++ app/src/main/java/com/zooop/zooop_android/models/UserDbHelper.java
@@ -1,0 +1,27 @@
+package com.zooop.zooop_android.models;
+
+import android.content.Context;
+import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteOpenHelper;
+
+/**
+ * Created by harpreet on 22/03/16.
+ */
+
+public class UserDbHelper extends SQLiteOpenHelper {
+
+    public UserDbHelper(Context context) {
+        super(context, UserContract.DATABASE_NAME, null, UserContract.DATABASE_VERSION);
+    }
+
+    @Override
+    public void onCreate(SQLiteDatabase db) {
+        db.execSQL(UserContract.UserEntry.SQL_CREATE_ENTRIES);
+    }
+
+    @Override
+    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        db.execSQL(UserContract.UserEntry.SQL_DELETE_ENTRIES);
+        onCreate(db);
+    }
+}
